make modules config immutable
